Refactored static code into components
I unintentionally made a much larger commit than I had originally planned, but I'm where I wanted to be.
Library contains the state for all of the books that belong to a bookshelf and I think I'm going to implement user-defined bookshelf names as a little project bonus. Next up is implementing loading the data from the API call.